+++ packages/integration-test/requiredDefault.test.cjs
@@ -1,6 +1,8 @@
 const withResolvers = require('core-js-pure/full/promise/with-resolvers');
 
 const {
+  asyncGeneratorWithLastValue,
+  generatorWithLastValue,
   iterableAt,
   iterableConcat,
   iterableEntries,
@@ -30,6 +32,38 @@ const {
   readerValues,
   SymbolObservable
 } = require('iter-fest');
+
+test('asyncGeneratorWithLastValue should work', async () => {
+  const asyncGenerator = asyncGeneratorWithLastValue(
+    (async function* () {
+      yield 1;
+
+      return 'end';
+    })()
+  );
+
+  for await (const value of asyncGenerator) {
+    expect(value).toBe(1);
+  }
+
+  expect(asyncGenerator.lastValue()).toEqual('end');
+});
+
+test('generatorWithLastValue should work', () => {
+  const generator = generatorWithLastValue(
+    (function* () {
+      yield 1;
+
+      return 'end';
+    })()
+  );
+
+  for (const value of generator) {
+    expect(value).toBe(1);
+  }
+
+  expect(generator.lastValue()).toEqual('end');
+});
 
 test('iterableAt should work', () => expect(iterableAt([1, 2, 3].values(), 1)).toBe(2));