+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `readerValues` in PR [#12](https://github.com/compulim/iter-fest/pull/12) and [#14](https://github.com/compulim/iter-fest/pull/14)
 - Added `observableSubscribeAsReadable` in PR [#13](https://github.com/compulim/iter-fest/pull/13)
 - Added `iterableGetReadable` in PR [#15](https://github.com/compulim/iter-fest/pull/15)
+- Added `generatorWithLastValue`/`asyncGeneratorWithLastValue` in PR [#XX](https://github.com/compulim/iter-fest/pull/XX)
 
 ### Changed